Did some conversion. Tweaks.

+++ 3D Graph Viewer/AugmentedRealityFiducialMarkerViewController.swift
@@ -27,8 +27,8 @@ class AugmentedRealityFiducialMarkerViewController: UIViewController, ARSCNViewD
     var originNode: SCNNode!
     var pickedImage: CIImage!
     var maxPointRadius: CGFloat = 0
-    var unitMeasure: Float = 10
-    var maxIndex: Float = 0
+    var unitMeasure: Double = 10
+    var maxIndex: Double = 0
     var shouldPlanesBeShown = true
     var shouldAxesLabelsBeShown = true
     
@@ -91,11 +91,11 @@ class AugmentedRealityFiducialMarkerViewController: UIViewController, ARSCNViewD
     private func placeScatterplotAt(position: simd_float4x4){
         var i = 0 // Variable used to identify sequentially a sphere inside the array
         for point in pointsToPlot{
-            let sphere = SCNSphere(radius: CGFloat(Float(point.sizeCoefficient) ?? 0.03))
+            let sphere = SCNSphere(radius: CGFloat(Double(point.sizeCoefficient) ?? 0.03))
             if(sphere.radius > maxPointRadius){
                 maxPointRadius = sphere.radius
             }
-            let tempMax = max(max(Float(point.xCoordinate)!, Float(point.yCoordinate)!), Float(point.zCoordinate)!)
+            let tempMax = max(max(Double(point.xCoordinate)!, Double(point.yCoordinate)!), Double(point.zCoordinate)!)
             if tempMax > maxIndex{
                 maxIndex = tempMax
             }
@@ -103,8 +103,8 @@ class AugmentedRealityFiducialMarkerViewController: UIViewController, ARSCNViewD
                 showLabels()
             }
             let sphereNode = SCNNode(geometry: sphere)
-            sphere.firstMaterial?.diffuse.contents = UIColor(red: CGFloat(Float(point.rColour) ?? 5)/255, green: CGFloat(Float(point.gColour) ?? 52)/255, blue: CGFloat(Float(point.bColour) ?? 105)/255, alpha: 1)
-            sphereNode.position = SCNVector3(Float(point.xCoordinate)!/unitMeasure, Float(point.yCoordinate)!/unitMeasure, Float(point.zCoordinate)!/unitMeasure)
+            sphere.firstMaterial?.diffuse.contents = UIColor(red: CGFloat(Double(point.rColour) ?? 5)/255, green: CGFloat(Double(point.gColour) ?? 52)/255, blue: CGFloat(Double(point.bColour) ?? 105)/255, alpha: 1)
+            sphereNode.position = SCNVector3(Double(point.xCoordinate)!/unitMeasure, Double(point.yCoordinate)!/unitMeasure, Double(point.zCoordinate)!/unitMeasure)
             let coordX = "\(sphereNode.position.x)"
             let coordY = "\(sphereNode.position.y)"
             let coordZ = "\(sphereNode.position.z)"
@@ -170,7 +170,7 @@ class AugmentedRealityFiducialMarkerViewController: UIViewController, ARSCNViewD
                     let textNode = SCNNode(geometry: textToShow)
                     textNode.name = "Info"
                     textNode.geometry?.firstMaterial?.diffuse.contents = UIColor(displayP3Red: 0/255, green: 0/255, blue: 0/255, alpha: 1) // UIColor needs values between 0 and 1 so the value is divided by 255
-                    textNode.position = SCNVector3(((tappedNode?.position.x)! + Float(maxPointRadius)), (tappedNode?.position.y)!, (tappedNode?.position.z)!)
+                    textNode.position = SCNVector3(Float((Double((tappedNode?.position.x)!)) + Double(maxPointRadius)), (tappedNode?.position.y)!, (tappedNode?.position.z)!)
                     textNode.scale = SCNVector3(0.002,0.002,0.002)
                     augmentedRealityFiducialMarkerScatterplot.scene.rootNode.addChildNode(textNode)
                 }
@@ -190,15 +190,15 @@ class AugmentedRealityFiducialMarkerViewController: UIViewController, ARSCNViewD
         for label in 0...(Int(maxIndex) + 1) {
             let labelToShow = SCNText(string: String(label), extrusionDepth: CGFloat(1))
             let labelNodeX = SCNNode(geometry: labelToShow)
-            labelNodeX.position = SCNVector3(Float(label)/unitMeasure, 0, 0)
+            labelNodeX.position = SCNVector3(Double(label)/unitMeasure, 0, 0)
             labelNodeX.geometry?.firstMaterial?.diffuse.contents = UIColor(displayP3Red: 0/255, green: 0/255, blue: 0/255, alpha: 1)
             labelNodeX.scale = SCNVector3(0.002,0.002,0.002)
             let labelNodeY = SCNNode(geometry: labelToShow)
-            labelNodeY.position = SCNVector3(0, Float(label)/unitMeasure, 0)
+            labelNodeY.position = SCNVector3(0, Double(label)/unitMeasure, 0)
             labelNodeY.geometry?.firstMaterial?.diffuse.contents = UIColor(displayP3Red: 0/255, green: 0/255, blue: 0/255, alpha: 1)
             labelNodeY.scale = SCNVector3(0.002,0.002,0.002)
             let labelNodeZ = SCNNode(geometry: labelToShow)
-            labelNodeZ.position = SCNVector3(0, 0, Float(label)/unitMeasure)
+            labelNodeZ.position = SCNVector3(0, 0, Double(label)/unitMeasure)
             labelNodeZ.geometry?.firstMaterial?.diffuse.contents = UIColor(displayP3Red: 0/255, green: 0/255, blue: 0/255, alpha: 1)
             labelNodeZ.scale = SCNVector3(0.002,0.002,0.002)
             augmentedRealityFiducialMarkerScatterplot.scene.rootNode.addChildNode(labelNodeX)
